Spited the classes into separate files
