+++ src/Tiles.py
@@ -1,0 +1,5 @@
+T = "🌳" # Tree
+P = "🦊" # Prey
+H = "👨" # Hunter
+M = "🗻" # Mountain
+F = "🌫️" # Fog
